Fix 3D plotting with matplotlib (#1158)
* Fix 3D plotting with matplotlib

* flake8

* Moved boundary_mesh_constructor to plottting

* Revert __all__ change

* update imports in plotting module
