+++ python/test/unit/mesh/test_surface_mesh.py
@@ -8,38 +8,11 @@ import dolfinx
 import dolfinx.cpp.mesh as cmesh
 import dolfinx.fem
 import dolfinx.io
-import dolfinx.mesh
+import dolfinx.plotting
 import numpy as np
 import pytest
 import ufl
 from mpi4py import MPI
-
-
-def create_boundary_mesh(mesh, comm, orient=False):
-    """
-    Create a mesh consisting of all exterior facets of a mesh
-    Input:
-      mesh   - The mesh
-      comm   - The MPI communicator
-      orient - Boolean flag for reorientation of facets to have
-               consistent outwards-pointing normal (default: True)
-    Output:
-      bmesh - The boundary mesh
-      bmesh_to_geometry - Map from cells of the boundary mesh
-                          to the geometry of the original mesh
-    """
-    ext_facets = cmesh.exterior_facet_indices(mesh)
-    boundary_geometry = cmesh.entities_to_geometry(
-        mesh, mesh.topology.dim - 1, ext_facets, orient)
-    facet_type = dolfinx.cpp.mesh.to_string(cmesh.cell_entity_type(
-        mesh.topology.cell_type, mesh.topology.dim - 1))
-    facet_cell = ufl.Cell(facet_type,
-                          geometric_dimension=mesh.geometry.dim)
-    degree = mesh.ufl_domain().ufl_coordinate_element().degree()
-    ufl_domain = ufl.Mesh(ufl.VectorElement("Lagrange", facet_cell, degree))
-    bmesh = dolfinx.mesh.create_mesh(
-        comm, boundary_geometry, mesh.geometry.x, ufl_domain)
-    return bmesh, boundary_geometry
 
 
 @pytest.mark.parametrize("celltype",
@@ -52,7 +25,7 @@ def test_b_mesh_mapping(celltype):
     """
     mesh = dolfinx.UnitCubeMesh(MPI.COMM_WORLD, 2, 2, 2, cell_type=celltype)
 
-    b_mesh, bndry_to_mesh = create_boundary_mesh(mesh, MPI.COMM_SELF)
+    b_mesh, bndry_to_mesh = dolfinx.plotting.create_boundary_mesh(mesh, MPI.COMM_SELF)
 
     # Compute map from boundary mesh topology to boundary mesh geometry
     b_mesh.topology.create_connectivity(
@@ -88,7 +61,7 @@ def test_b_mesh_orientation(celltype):
                             np.array([0.5, 0.5, 0.5])],
                            [2, 2, 2], cell_type=celltype)
 
-    b_mesh, bndry_to_mesh = create_boundary_mesh(mesh, MPI.COMM_SELF, True)
+    b_mesh, bndry_to_mesh = dolfinx.plotting.create_boundary_mesh(mesh, MPI.COMM_SELF, True)
     bdim = b_mesh.topology.dim
     b_mesh.topology.create_connectivity(bdim, bdim - 1)
     b_mesh.topology.create_connectivity(bdim - 1, bdim)
